refactor(component): ♻️  amenity section at target card as component
